Make tv_nsec long

+++ source/vox/lib/system/linux.d
@@ -10,8 +10,8 @@ import vox.lib.types;
 enum CLOCK_MONOTONIC = 1;
 
 struct timespec {
-	long  tv_sec;
-	int   tv_nsec;
+	long tv_sec;
+	long tv_nsec;
 }
 
 int clock_gettime(int clockid, timespec* tp) {
